+++ packages/isbt/test/lint.test.ts
@@ -1,0 +1,11 @@
+import { runIsbtCommandOnFixture } from './runIsbtCommandOnFixture'
+
+describe('lint', () => {
+  it('passes when missing files', async () => {
+    await runIsbtCommandOnFixture('workspace-1', 'lint . --passWithNoTests')
+  })
+
+  it('passes when missing config', async () => {
+    await runIsbtCommandOnFixture('workspace-2', 'lint . --passWithNoTests')
+  })
+})
